# fix: perf improvements to ring-buffer