Before:
```
Platform: Darwin (23.5.0)
CPU: Apple M1
Node.js v18.19.1
Date: 7/18/2024, 1:28:44 PM
┌────────────────────────┬──────────────────┬──────────────────┬──────────────────┬────────────────────┐
│        (index)         │   1 small file   │   1 large file   │ 100 small files  │   5 large files    │
├────────────────────────┼──────────────────┼──────────────────┼──────────────────┼────────────────────┤
│ fetch-multipart-parser │ '0.10 ms ± 0.55' │ '7.24 ms ± 0.45' │ '0.35 ms ± 0.20' │ '81.04 ms ± 14.45' │
└────────────────────────┴──────────────────┴──────────────────┴──────────────────┴────────────────────┘
```
After:
```
Platform: Darwin (23.5.0)
CPU: Apple M1
Node.js v18.19.1
Date: 7/18/2024, 1:29:11 PM
┌────────────────────────┬──────────────────┬──────────────────┬──────────────────┬───────────────────┐
│        (index)         │   1 small file   │   1 large file   │ 100 small files  │   5 large files   │
├────────────────────────┼──────────────────┼──────────────────┼──────────────────┼───────────────────┤
│ fetch-multipart-parser │ '0.08 ms ± 0.44' │ '6.56 ms ± 0.40' │ '0.29 ms ± 0.06' │ '69.38 ms ± 4.16' │
└────────────────────────┴──────────────────┴──────────────────┴──────────────────┴───────────────────┘
```